Moved IG to a social app
