+++ ads/urls.py
@@ -4,7 +4,5 @@ from . import views
 app_name = 'ads'
 
 urlpatterns = [
-    path('update/', views.igUrl, name='igurl'),
     path('', views.adsIndex, name='index'),
-
 ]
