+++ core/urls.py
@@ -32,6 +32,7 @@ urlpatterns = [
     path('events/', include('events.urls')),
     path('guides/', include('staticpages.urls')),
     path('ads/',include('ads.urls')),
+    path('social/',include('social.urls')),
     path('admin/', admin.site.urls),
 ]
 urlpatterns += static(settings.MEDIA_URL, document_root=settings.MEDIA_ROOT)
